Move files to middleware
