+++ src/lib/Repository/index.ts
@@ -1,2 +1,1 @@
 export * from './AbstractMongoDBPayloadRepository'
-export * from './Archive'
